refactor: 로또 자동번호 생성 로직 상위 레벨로 이동
